backend/pyproject.toml
[project]
name = "ai-chatbot-backend"
version = "0.1.0"
description = "AI Customer Service Chatbot Backend (Django + Ninja + LlamaIndex)"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    # ---------- 核心框架 (Core) ----------
    "django>=6.0.1",                # Web 框架基石 (自带 ORM, Admin 后台)
    "django-simpleui>=2026.1.13",    # 现代化 Admin UI (Element UI 风格，中文友好)
    "django-ninja>=1.5.0,<2.0",     # API 扩展框架 (放宽版本以兼容 JWT)
    
    # ---------- AI 引擎 (RAG Engine) ----------
    "llama-index>=0.14.12",
    "llama-index-llms-openai-like>=0.3.0",
    "llama-index-embeddings-openai>=0.3.0",       # OpenAI Embedding (备用)
    "llama-index-embeddings-huggingface>=0.5.0",  # 本地 HuggingFace Embedding
    "llama-index-readers-file>=0.4.0",            # 文件解析器
    "llama-index-vector-stores-postgres>=0.4.0",  # Postgres 向量存储
    "pypdf>=5.0.0",                               # PDF 解析库
    "pdfplumber>=0.11.0",                         # 更稳定的 PDF 解析（中文友好）
    "docx2txt>=0.8",                              # DOCX 解析
    "python-pptx>=0.6.21",                        # PPTX 解析
    "html2text>=2020.1.16",                       # HTML/EPUB 文本提取
    "openpyxl>=3.1.2",                            # Excel 解析
    "pandas>=2.2.0",                              # 表格数据处理
    "llama-index-tools-mcp>=0.3.0",               # MCP 工具集成

    # ---------- 数据库与向量 (DB & Vector) ----------
    "psycopg[binary]>=3.2.0",       # PostgreSQL 数据库驱动 (连接数据库必用)
    # "pgvector",                   # (已由 llama-index-vector-stores-postgres 自动安装, 无需手动指定)

    # ---------- Web 服务增强 ----------
    "django-cors-headers>=4.3.1",   # 跨域资源共享 (允许前端跨域访问 API)
    "requests>=2.31.0",             # HTTP 请求库 (调用外部 API)

    # ---------- 用户认证 (Auth) ----------
    "PyJWT>=2.8.0",                 # Python 官方 JWT 库 (稳定可靠)

    # ---------- 环境与服务 (Server) ----------
    "python-dotenv>=1.0.0",         # 环境变量加载工具 (读取 .env 文件中的 keys)
    "gunicorn>=23.0.0",             # 生产级 WSGI 服务器 (进程管理器)
    "uvicorn>=0.40.0",              # 高性能 ASGI 服务器 (处理异步并发请求)
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
